refactor: JSDOC기반 주석 추가
